+++ lib/mixins/use_sprite_animation.dart
@@ -123,7 +123,7 @@ mixin UseSpriteAnimation on GameComponent {
   int get animationCurrentIndex => _animationRender?.currentIndex ?? 0;
   bool get isPaused => _animationRender?.isPaused ?? false;
 
-  void showAnimationStroke(Color color, double width) {
+  void showAnimationStroke(Color color, double width,{Vector2? offset}) {
     if (_strockePaint != null &&
         _strokeWidth == width &&
         _strockePaint?.color == color) {
@@ -131,6 +131,9 @@ mixin UseSpriteAnimation on GameComponent {
     }
     _strokeWidth = width;
     _strokePosition = Vector2.all(-_strokeWidth);
+    if (offset != null) {
+      _strokePosition += offset;
+    }
     _strokeSize = Vector2.zero();
     _strockePaint = Paint()
       ..color = color