+++ lib/util/direction_animations/simple_direction_animation.dart
@@ -518,7 +518,7 @@ class SimpleDirectionAnimation {
     }
   }
 
-  void showStroke(Color color, double width) {
+  void showStroke(Color color, double width, {Vector2? offset}) {
     if (_strockePaint != null &&
         _strokeWidth == width &&
         _strockePaint?.color == color) {
@@ -526,6 +526,9 @@ class SimpleDirectionAnimation {
     }
     _strokeWidth = width;
     _strokePosition = Vector2.all(-_strokeWidth);
+    if (offset != null) {
+      _strokePosition += offset;
+    }
     _strokeSize = Vector2.zero();
     _strockePaint = Paint()
       ..color = color