+++ lib/npc/enemy/platform_enemy.dart
@@ -21,16 +21,13 @@ class PlatformEnemy extends SimpleEnemy
             runLeft: animation.runLeft,
             others: {
               if (animation.jump?.jumpUpRight != null)
-                JumpAnimationsEnum.jumpUpRight:
-                    animation.jump!.jumpUpRight,
+                JumpAnimationsEnum.jumpUpRight: animation.jump!.jumpUpRight,
               if (animation.jump?.jumpUpLeft != null)
                 JumpAnimationsEnum.jumpUpLeft: animation.jump!.jumpUpLeft!,
               if (animation.jump?.jumpDownRight != null)
-                JumpAnimationsEnum.jumpDownRight:
-                    animation.jump!.jumpDownRight,
+                JumpAnimationsEnum.jumpDownRight: animation.jump!.jumpDownRight,
               if (animation.jump?.jumpDownLeft != null)
-                JumpAnimationsEnum.jumpDownLeft:
-                    animation.jump!.jumpDownLeft!,
+                JumpAnimationsEnum.jumpDownLeft: animation.jump!.jumpDownLeft!,
               ...animation.others ?? {},
             },
           ),