+++ lib/util/extensions/movement_extensions.dart
@@ -2,6 +2,8 @@ import 'dart:math';
 
 import 'package:bonfire/bonfire.dart';
 
+enum MovementAxis { horizontal, vertical, withoutDiagonal, all }
+
 extension MovementExtensions on Movement {
   /// This method move this component to target
   /// Need use Movement mixin.
@@ -11,6 +13,7 @@ extension MovementExtensions on Movement {
     required T target,
     Function? close,
     double margin = 4,
+    MovementAxis movementAxis = MovementAxis.all,
   }) {
     Rect rectPlayerCollision = target.rectCollision.inflate(margin);
 
@@ -21,12 +24,23 @@ extension MovementExtensions on Movement {
     }
 
     double radAngle = getAngleFromTarget(target);
-    final direct = BonfireUtil.getDirectionFromAngle(radAngle);
-    if (canMove(direct, ignoreHitboxes: target.shapeHitboxes)) {
-      moveFromDirection(direct);
+    Direction directionToMove = BonfireUtil.getDirectionFromAngle(radAngle);
+    final newDirectionToMove = _checkRestrictAxis(
+      directionToMove,
+      movementAxis,
+    );
+    if (newDirectionToMove != null) {
+      directionToMove = newDirectionToMove;
+    } else {
+      stopMove();
+      return false;
+    }
+
+    if (canMove(directionToMove, ignoreHitboxes: target.shapeHitboxes)) {
+      moveFromDirection(directionToMove);
       return true;
     } else {
-      switch (direct) {
+      switch (directionToMove) {
         case Direction.right:
         case Direction.left:
         case Direction.up:
@@ -212,5 +226,80 @@ extension MovementExtensions on Movement {
         }
       }
     }
+  }
+
+  Direction? _checkRestrictAxis(
+    Direction directionToMove,
+    MovementAxis moveAxis,
+  ) {
+    if (moveAxis == MovementAxis.all) {
+      return directionToMove;
+    }
+    switch (directionToMove) {
+      case Direction.upLeft:
+        switch (moveAxis) {
+          case MovementAxis.horizontal:
+          case MovementAxis.withoutDiagonal:
+            return Direction.left;
+          case MovementAxis.vertical:
+            return Direction.up;
+
+          default:
+        }
+      case Direction.upRight:
+        switch (moveAxis) {
+          case MovementAxis.horizontal:
+          case MovementAxis.withoutDiagonal:
+            return Direction.right;
+          case MovementAxis.vertical:
+            return Direction.up;
+
+          default:
+        }
+      case Direction.downLeft:
+        switch (moveAxis) {
+          case MovementAxis.horizontal:
+          case MovementAxis.withoutDiagonal:
+            return Direction.left;
+          case MovementAxis.vertical:
+            return Direction.down;
+
+          default:
+        }
+      case Direction.downRight:
+        switch (moveAxis) {
+          case MovementAxis.horizontal:
+          case MovementAxis.withoutDiagonal:
+            return Direction.right;
+          case MovementAxis.vertical:
+            return Direction.down;
+          default:
+        }
+      case Direction.left:
+        if (moveAxis == MovementAxis.vertical) {
+          return null;
+        } else {
+          return directionToMove;
+        }
+      case Direction.right:
+        if (moveAxis == MovementAxis.vertical) {
+          return null;
+        } else {
+          return directionToMove;
+        }
+      case Direction.up:
+        if (moveAxis == MovementAxis.horizontal) {
+          return null;
+        } else {
+          return directionToMove;
+        }
+      case Direction.down:
+        if (moveAxis == MovementAxis.horizontal) {
+          return null;
+        } else {
+          return directionToMove;
+        }
+    }
+    return null;
   }
 }