+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Adds bool `paused` in `BonfireGameInterface`.
 - `SimpleDirectionAnimation.playOther` improvements.
 - `TileRecognizer` improvements.
+- Render priority improvements.
 
 ## 3.0.5
 - Update `a_star_algorithm`